+++ tests/Unit/TestBuilderTest.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2021 Cloud Creativity Limited
+ * Copyright 2022 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,9 +19,11 @@ declare(strict_types=1);
 
 namespace LaravelJsonApi\Testing\Tests\Unit;
 
+use Illuminate\Contracts\Routing\UrlRoutable;
 use Illuminate\Foundation\Testing\Concerns\MakesHttpRequests;
 use Illuminate\Http\Response;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 use Illuminate\Testing\TestResponse as IlluminateTestResponse;
 use LaravelJsonApi\Testing\TestBuilder;
@@ -445,6 +447,7 @@ class TestBuilderTest extends TestCase
         $query = [
             'filter' => [
                 'published' => 'true',
+                'foo' => 'bar',
             ],
         ];
 
@@ -458,7 +461,233 @@ class TestBuilderTest extends TestCase
 
         $response = $this->builder
             ->expects('posts')
-            ->filter(['published' => 'true'])
+            ->filter(['published' => 'true', 'foo' => 'bar'])
+            ->get('/api/v1/posts');
+
+        $this->assertEquals(new TestResponse($this->response, 'posts'), $response);
+    }
+
+    /**
+     * @return array
+     */
+    public function idProvider(): array
+    {
+        return [
+            'int' => [static fn() => 1],
+            'string' => [static fn() => '1'],
+            'model' => [static function ($test) {
+                $mock = $test->createMock(UrlRoutable::class);
+                $mock->method('getRouteKey')->willReturn(1);
+                return $mock;
+            }],
+        ];
+    }
+
+    /**
+     * @param \Closure $scenario
+     * @return void
+     * @dataProvider idProvider
+     */
+    public function testFilterId(\Closure $scenario): void
+    {
+        $value = $scenario($this);
+
+        $headers = [
+            'Accept' => 'application/vnd.api+json',
+            'CONTENT_TYPE' => 'application/vnd.api+json',
+        ];
+
+        $query = [
+            'filter' => [
+                'published' => 'true',
+                'author' => '1',
+            ],
+        ];
+
+        $expected = '/api/v1/posts?' . Arr::query($query);
+
+        $this->mock
+            ->expects($this->once())
+            ->method('json')
+            ->with('GET', $expected, [], $headers)
+            ->willReturn(new IlluminateTestResponse($this->response));
+
+        $response = $this->builder
+            ->expects('posts')
+            ->filter(['published' => 'true', 'author' => $value])
+            ->get('/api/v1/posts');
+
+        $this->assertEquals(new TestResponse($this->response, 'posts'), $response);
+    }
+
+    /**
+     * @param \Closure $scenario
+     * @return void
+     * @dataProvider idProvider
+     */
+    public function testFilterIdInQuery(\Closure $scenario): void
+    {
+        $value = $scenario($this);
+
+        $headers = [
+            'Accept' => 'application/vnd.api+json',
+            'CONTENT_TYPE' => 'application/vnd.api+json',
+        ];
+
+        $query = [
+            'filter' => [
+                'published' => 'true',
+                'author' => '1',
+            ],
+        ];
+
+        $expected = '/api/v1/posts?' . Arr::query($query);
+
+        $this->mock
+            ->expects($this->once())
+            ->method('json')
+            ->with('GET', $expected, [], $headers)
+            ->willReturn(new IlluminateTestResponse($this->response));
+
+        $response = $this->builder
+            ->expects('posts')
+            ->query(['filter' => ['published' => 'true', 'author' => $value]])
+            ->get('/api/v1/posts');
+
+        $this->assertEquals(new TestResponse($this->response, 'posts'), $response);
+    }
+
+    /**
+     * @return array
+     */
+    public function idsProvider(): array
+    {
+        return [
+            'integers' => [static fn() => [1, 2, 3]],
+            'strings' => [static fn() => ['1', '2', '3']],
+            'models' => [
+                static function ($test) {
+                    $model1 = $test->createMock(UrlRoutable::class);
+                    $model1->method('getRouteKey')->willReturn(1);
+
+                    $model2 = $test->createMock(UrlRoutable::class);
+                    $model2->method('getRouteKey')->willReturn(2);
+
+                    $model3 = $test->createMock(UrlRoutable::class);
+                    $model3->method('getRouteKey')->willReturn(3);
+
+                    return [$model1, $model2, $model3];
+                },
+            ],
+        ];
+    }
+
+    /**
+     * @param \Closure $scenario
+     * @return void
+     * @dataProvider idsProvider
+     */
+    public function testFilterIds(\Closure $scenario): void
+    {
+        $values = $scenario($this);
+
+        $headers = [
+            'Accept' => 'application/vnd.api+json',
+            'CONTENT_TYPE' => 'application/vnd.api+json',
+        ];
+
+        $query = [
+            'filter' => [
+                'published' => 'true',
+                'id' => ['1', '2', '3'],
+            ],
+        ];
+
+        $expected = '/api/v1/posts?' . Arr::query($query);
+
+        $this->mock
+            ->expects($this->once())
+            ->method('json')
+            ->with('GET', $expected, [], $headers)
+            ->willReturn(new IlluminateTestResponse($this->response));
+
+        $response = $this->builder
+            ->expects('posts')
+            ->filter(['published' => 'true', 'id' => $values])
+            ->get('/api/v1/posts');
+
+        $this->assertEquals(new TestResponse($this->response, 'posts'), $response);
+    }
+
+    /**
+     * @param \Closure $scenario
+     * @return void
+     * @dataProvider idsProvider
+     */
+    public function testFilterIdsWithCollection(\Closure $scenario): void
+    {
+        $values = $scenario($this);
+
+        $headers = [
+            'Accept' => 'application/vnd.api+json',
+            'CONTENT_TYPE' => 'application/vnd.api+json',
+        ];
+
+        $query = [
+            'filter' => [
+                'published' => 'true',
+                'ids' => ['1', '2', '3'],
+            ],
+        ];
+
+        $expected = '/api/v1/posts?' . Arr::query($query);
+
+        $this->mock
+            ->expects($this->once())
+            ->method('json')
+            ->with('GET', $expected, [], $headers)
+            ->willReturn(new IlluminateTestResponse($this->response));
+
+        $response = $this->builder
+            ->expects('posts')
+            ->filter(['published' => 'true', 'ids' => new Collection($values)])
+            ->get('/api/v1/posts');
+
+        $this->assertEquals(new TestResponse($this->response, 'posts'), $response);
+    }
+
+    /**
+     * @param \Closure $scenario
+     * @return void
+     * @dataProvider idsProvider
+     */
+    public function testFilterIdsInQuery(\Closure $scenario): void
+    {
+        $values = $scenario($this);
+
+        $headers = [
+            'Accept' => 'application/vnd.api+json',
+            'CONTENT_TYPE' => 'application/vnd.api+json',
+        ];
+
+        $query = [
+            'filter' => [
+                'published' => 'true',
+                'ids' => ['1', '2', '3'],
+            ],
+        ];
+
+        $expected = '/api/v1/posts?' . Arr::query($query);
+
+        $this->mock
+            ->expects($this->once())
+            ->method('json')
+            ->with('GET', $expected, [], $headers)
+            ->willReturn(new IlluminateTestResponse($this->response));
+
+        $response = $this->builder
+            ->expects('posts')
+            ->query(['filter' => ['published' => 'true', 'ids' => $values]])
             ->get('/api/v1/posts');
 
         $this->assertEquals(new TestResponse($this->response, 'posts'), $response);
@@ -486,6 +715,110 @@ class TestBuilderTest extends TestCase
         $response = $this->builder
             ->expects('posts')
             ->sort('-publishedAt', 'title')
+            ->get('/api/v1/posts');
+
+        $this->assertEquals(new TestResponse($this->response, 'posts'), $response);
+    }
+
+    public function testPage(): void
+    {
+        $headers = [
+            'Accept' => 'application/vnd.api+json',
+            'CONTENT_TYPE' => 'application/vnd.api+json',
+        ];
+
+        $query = [
+            'page' => [
+                'number' => '2',
+                'size' => '10',
+            ],
+        ];
+
+        $expected = '/api/v1/posts?' . Arr::query($query);
+
+        $this->mock
+            ->expects($this->once())
+            ->method('json')
+            ->with('GET', $expected, [], $headers)
+            ->willReturn(new IlluminateTestResponse($this->response));
+
+        $response = $this->builder
+            ->expects('posts')
+            ->page(['number' => '2', 'size' => '10'])
+            ->get('/api/v1/posts');
+
+        $this->assertEquals(new TestResponse($this->response, 'posts'), $response);
+    }
+
+    /**
+     * @param \Closure $scenario
+     * @return void
+     * @dataProvider idProvider
+     */
+    public function testPageId(\Closure $scenario): void
+    {
+        $value = $scenario($this);
+
+        $headers = [
+            'Accept' => 'application/vnd.api+json',
+            'CONTENT_TYPE' => 'application/vnd.api+json',
+        ];
+
+        $query = [
+            'page' => [
+                'after' => '1',
+                'size' => '10',
+            ],
+        ];
+
+        $expected = '/api/v1/posts?' . Arr::query($query);
+
+        $this->mock
+            ->expects($this->once())
+            ->method('json')
+            ->with('GET', $expected, [], $headers)
+            ->willReturn(new IlluminateTestResponse($this->response));
+
+        $response = $this->builder
+            ->expects('posts')
+            ->page(['after' => $value, 'size' => '10'])
+            ->get('/api/v1/posts');
+
+        $this->assertEquals(new TestResponse($this->response, 'posts'), $response);
+    }
+
+    /**
+     * @param \Closure $scenario
+     * @return void
+     * @dataProvider idProvider
+     */
+    public function testPageIdInQuery(\Closure $scenario): void
+    {
+        $value = $scenario($this);
+
+        $headers = [
+            'Accept' => 'application/vnd.api+json',
+            'CONTENT_TYPE' => 'application/vnd.api+json',
+        ];
+
+        $query = [
+            'page' => [
+                'after' => '1',
+                'size' => '10',
+            ],
+        ];
+
+        $expected = '/api/v1/posts?' . Arr::query($query);
+
+        $this->mock
+            ->expects($this->once())
+            ->method('json')
+            ->with('GET', $expected, [], $headers)
+            ->willReturn(new IlluminateTestResponse($this->response));
+
+        $response = $this->builder
+            ->expects('posts')
+            ->query(['page' => ['after' => $value, 'size' => '10']])
             ->get('/api/v1/posts');
 
         $this->assertEquals(new TestResponse($this->response, 'posts'), $response);